[refactor] 상수 이넘화 (#4)
