+++ src/main/java/konkuk/chacall/global/common/security/constant/AuthParameters.java
@@ -6,6 +6,8 @@ import lombok.Getter;
 public enum AuthParameters {
     JWT_HEADER_KEY("Authorization"),
     JWT_PREFIX("Bearer "),
+    REDIRECT_URI_KEY("redirect_uri"),
+    REDIRECT_SESSION_KEY("oauth2_return_to"),
     KAKAO("kakao"),
     KAKAO_PROVIDER_ID_KEY("id"),
     JWT_ACCESS_TOKEN_KEY("userId"),
@@ -13,7 +15,6 @@ public enum AuthParameters {
     REDIRECT_ROLE_SELECT_URL("/index.html"),
     REDIRECT_MEMBER_HOME_URL("/member"),
     REDIRECT_OWNER_HOME_URL("/owner"),
-
     ;
 
     private final String value;
